Fixed majors issues when adding direct t solution
Added the option to get the t emission directly with a seperate emission option for this. Also fixed the sudakov basis such that the betas are calculated iteratively instead of just with the alphas. Also fixed too many gluon emissions with a missing 2pi term in the g -> gg splitting function